requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.2
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
